GuestBookServer/serverapp/frontend1/requirements.txt
Django==3.1.5
django-extensions==3.1.0
# gunicorn==20.0.4 # app engine instructions say not to install this unless you have your own entrypoint
psycopg2-binary==2.8.6
# pillow
numpy==1.19.5
django-allauth==0.44.0
ffmpeg-python==0.2.0
segno==1.3.1
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-cloud-tasks==2.1.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
protobuf==3.14.0
